Support macOS 14
macOS version 14 removed the `/usr/local/var/` directory. `/run/` and
`/var/run/` are defined by the Filesystem Hierarchy Standard for the
purpose of UNIX-domain sockets [1]. However, macOS 14 does not include
`/run/`, and `/var/run/` is not world-writable on that platform. In
order to support macOS 14, use the `/tmp/` directory, instead. Extend
the configuration of the project's continuous integration environment to
include macOS 14.

This commit has been cherry-picked from the "aria-at-automation-driver"
project, where it was previously reviewed and approved [2].

[1] https://refspecs.linuxfoundation.org/FHS_3.0/index.html
[2] https://github.com/w3c/aria-at-automation-driver/pull/55