Extend timeout value for tests in CI
The CI environment currently used by the project has been observed to
require between 1497ms and 10398ms to start a WebDriver session with
Firefox. Increase the upper limit to accommodate the worst-case
scenario, but do so only when the test suite is known to be running in
the continuous integration environment.